[(fixes)#102586948] modified sign-in-feature
