+++ codango/account/templates/account/base.html
@@ -1,0 +1,41 @@
+{% load staticfiles %}
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <link href="/static/css/bootstrap.min.css" rel="stylesheet">
+    <link href="/static/css/app.css" type="text/css" rel="stylesheet"/>
+    <title>{% block title %}Codango{% endblock title %}</title>
+  </head>
+
+  <body>
+    <!--header-->
+    <div id="header" class ="container-fluid">
+	    <div class="container">
+	      <div class="row">
+          {% block header %}
+	        <h1 class="col-md-8 col-xs-8">Codango</h1>
+	        <a class="col-md-1 col-xs-2" href = "#">The Team</a>
+	        <a class="col-md-3 col-xs-2" href = "#">Contact us</a> 
+          {% endblock header %}
+	     </div>
+      </div>
+    </div>
+    
+    <!--page content-->
+    <div class="container">
+      {% block content %}{% endblock content %}
+      
+    </div>
+
+    <!--footer-->
+    <div id="footer" class="container-fluid">
+      <div class="panel panel-default">
+          <h4 class="pull-right"><small>&copy 2015 CODANGO</small></h4>
+      </div>
+    </div>
+
+    <script src="/static/js/jquery-1.11.3.min.js"></script>
+    <script src="/static/js/bootstrap.min.js"></script>
+
+  </body>
+</html>
