+++ codango/account/templates/account/home.html
@@ -1,10 +1,12 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Home</title>
-</head>
-<body>
+{% extends "account/base.html" %}
+
+{% block title %}Home{% endblock title %}
+
+{% block header %}
+	<h1>Codango</h1>
+{% endblock header %}
+
+{% block content %}
+  <h1>Codango Home</h1>
   <p>Logged in successfully</p>
-</body>
-</html>
+{% endblock content %}
