+++ codango/account/tests/test_views.py
@@ -1,42 +1,31 @@
-from django.test import TestCase, Client
-from django.contrib.auth.models import User
-from django.core.urlresolvers import resolve
+from django.contrib.staticfiles.testing import StaticLiveServerTestCase
+from selenium import webdriver
+from selenium.webdriver.common.keys import Keys
 
 
-class IndexViewTest(TestCase):
-
-    def setUp(self):
-        self.client = Client()
-        User.objects.create_user(
-            username='lade',
-            password='password'
-        )
-
-    def test_can_reach_index_page(self):
-        response = self.client.get('/')
-        self.assertEqual(response.status_code, 200)
-
-    def test_right_view_for_index_is_returned(self):
-        match = resolve('/')
-        self.assertEqual(match.url_name, 'index')
-
-    def test_can_login(self):
-        response = self.client.post('/', {
-            'username': 'lade',
-            'password': 'password'
-        })
-        self.assertEqual(response.status_code, 302)
-
-
-class HomeViewTest(TestCase):
+class IndexViewTest(StaticLiveServerTestCase):
+    fixtures = ['users.json']
 
     def setUp(self):
-        self.client = Client()
+        self.browser = webdriver.Firefox()
 
-    def test_can_reach_home_page(self):
-        response = self.client.get('/home')
-        self.assertEqual(response.status_code, 301)
+    def tearDown(self):
+        self.browser.quit()
 
-    def test_right_view_for_home_is_returned(self):
-        match = resolve('/home/')
-        self.assertEqual(match.url_name, 'home')
+    def test_can_reach_index_page_and_log_in(self):
+        self.browser.get(self.live_server_url)
+
+        body = self.browser.find_element_by_tag_name('body')
+        self.assertIn('Codango', body.text)
+
+        # testing username and password
+        username_field = self.browser.find_element_by_name('username')
+        username_field.send_keys('lade')
+
+        password_field = self.browser.find_element_by_name('password')
+        password_field.send_keys('password')
+        password_field.send_keys(Keys.RETURN)
+
+        # username and password accepted
+        body = self.browser.find_element_by_tag_name('body')
+        self.assertIn('Codango Home', body.text)
